Refactor: move page components into their own folder
